Generate Numeric List

generate_numeric_list generates a random list to simulate what vector embeddings look like

Usage

generate_numeric_list(decimals = 1, length)

Arguments

decimals

the number of decimals needed for each figure. Defaults to 1.

length

the length of the list you want to produce

Details

Generates a list of numerical values of a specific length. It can be used to simulate what vector embeddings look like.

Examples

generate_numeric_list(decimals=1, length=384)
#> [[1]]
#> [1] 0.1
#> 
#> [[2]]
#> [1] 0.8
#> 
#> [[3]]
#> [1] 0.6
#> 
#> [[4]]
#> [1] 0.2
#> 
#> [[5]]
#> [1] 0
#> 
#> [[6]]
#> [1] 0.5
#> 
#> [[7]]
#> [1] 0.5
#> 
#> [[8]]
#> [1] 0.3
#> 
#> [[9]]
#> [1] 0.7
#> 
#> [[10]]
#> [1] 0.8
#> 
#> [[11]]
#> [1] 0.9
#> 
#> [[12]]
#> [1] 0.2
#> 
#> [[13]]
#> [1] 0
#> 
#> [[14]]
#> [1] 0.3
#> 
#> [[15]]
#> [1] 0.4
#> 
#> [[16]]
#> [1] 0.2
#> 
#> [[17]]
#> [1] 0.4
#> 
#> [[18]]
#> [1] 0.1
#> 
#> [[19]]
#> [1] 0.4
#> 
#> [[20]]
#> [1] 1
#> 
#> [[21]]
#> [1] 0.3
#> 
#> [[22]]
#> [1] 0.7
#> 
#> [[23]]
#> [1] 0.7
#> 
#> [[24]]
#> [1] 0.2
#> 
#> [[25]]
#> [1] 1
#> 
#> [[26]]
#> [1] 0.7
#> 
#> [[27]]
#> [1] 0.1
#> 
#> [[28]]
#> [1] 0.5
#> 
#> [[29]]
#> [1] 0.7
#> 
#> [[30]]
#> [1] 0.7
#> 
#> [[31]]
#> [1] 0
#> 
#> [[32]]
#> [1] 0.2
#> 
#> [[33]]
#> [1] 0.3
#> 
#> [[34]]
#> [1] 0.6
#> 
#> [[35]]
#> [1] 0.5
#> 
#> [[36]]
#> [1] 0.4
#> 
#> [[37]]
#> [1] 0.7
#> 
#> [[38]]
#> [1] 0.9
#> 
#> [[39]]
#> [1] 0.2
#> 
#> [[40]]
#> [1] 0.2
#> 
#> [[41]]
#> [1] 0.7
#> 
#> [[42]]
#> [1] 0.5
#> 
#> [[43]]
#> [1] 0.6
#> 
#> [[44]]
#> [1] 0.7
#> 
#> [[45]]
#> [1] 0.1
#> 
#> [[46]]
#> [1] 0.8
#> 
#> [[47]]
#> [1] 0.8
#> 
#> [[48]]
#> [1] 1
#> 
#> [[49]]
#> [1] 1
#> 
#> [[50]]
#> [1] 0.4
#> 
#> [[51]]
#> [1] 0.5
#> 
#> [[52]]
#> [1] 0.3
#> 
#> [[53]]
#> [1] 0.2
#> 
#> [[54]]
#> [1] 0.5
#> 
#> [[55]]
#> [1] 0.5
#> 
#> [[56]]
#> [1] 0.8
#> 
#> [[57]]
#> [1] 0.2
#> 
#> [[58]]
#> [1] 0.7
#> 
#> [[59]]
#> [1] 0.1
#> 
#> [[60]]
#> [1] 0.4
#> 
#> [[61]]
#> [1] 0.8
#> 
#> [[62]]
#> [1] 0.3
#> 
#> [[63]]
#> [1] 0.6
#> 
#> [[64]]
#> [1] 0.3
#> 
#> [[65]]
#> [1] 0.6
#> 
#> [[66]]
#> [1] 0.2
#> 
#> [[67]]
#> [1] 0.9
#> 
#> [[68]]
#> [1] 0.5
#> 
#> [[69]]
#> [1] 0.5
#> 
#> [[70]]
#> [1] 0.3
#> 
#> [[71]]
#> [1] 0.4
#> 
#> [[72]]
#> [1] 0.4
#> 
#> [[73]]
#> [1] 0
#> 
#> [[74]]
#> [1] 0.5
#> 
#> [[75]]
#> [1] 0.4
#> 
#> [[76]]
#> [1] 0
#> 
#> [[77]]
#> [1] 0.4
#> 
#> [[78]]
#> [1] 0.6
#> 
#> [[79]]
#> [1] 0.9
#> 
#> [[80]]
#> [1] 0.4
#> 
#> [[81]]
#> [1] 0.5
#> 
#> [[82]]
#> [1] 0.6
#> 
#> [[83]]
#> [1] 0.3
#> 
#> [[84]]
#> [1] 0.3
#> 
#> [[85]]
#> [1] 0.5
#> 
#> [[86]]
#> [1] 0.9
#> 
#> [[87]]
#> [1] 0.4
#> 
#> [[88]]
#> [1] 0.2
#> 
#> [[89]]
#> [1] 0.7
#> 
#> [[90]]
#> [1] 0.1
#> 
#> [[91]]
#> [1] 1
#> 
#> [[92]]
#> [1] 0.1
#> 
#> [[93]]
#> [1] 0.5
#> 
#> [[94]]
#> [1] 0.8
#> 
#> [[95]]
#> [1] 0.7
#> 
#> [[96]]
#> [1] 0.2
#> 
#> [[97]]
#> [1] 0.5
#> 
#> [[98]]
#> [1] 0.8
#> 
#> [[99]]
#> [1] 0
#> 
#> [[100]]
#> [1] 0.5
#> 
#> [[101]]
#> [1] 0.8
#> 
#> [[102]]
#> [1] 0.8
#> 
#> [[103]]
#> [1] 0.4
#> 
#> [[104]]
#> [1] 0.2
#> 
#> [[105]]
#> [1] 0.4
#> 
#> [[106]]
#> [1] 0.7
#> 
#> [[107]]
#> [1] 0.5
#> 
#> [[108]]
#> [1] 0.7
#> 
#> [[109]]
#> [1] 0.5
#> 
#> [[110]]
#> [1] 0.8
#> 
#> [[111]]
#> [1] 0.7
#> 
#> [[112]]
#> [1] 0.9
#> 
#> [[113]]
#> [1] 0
#> 
#> [[114]]
#> [1] 0.9
#> 
#> [[115]]
#> [1] 1
#> 
#> [[116]]
#> [1] 0.5
#> 
#> [[117]]
#> [1] 0.4
#> 
#> [[118]]
#> [1] 0.8
#> 
#> [[119]]
#> [1] 0.6
#> 
#> [[120]]
#> [1] 0.6
#> 
#> [[121]]
#> [1] 0.9
#> 
#> [[122]]
#> [1] 0.6
#> 
#> [[123]]
#> [1] 0.3
#> 
#> [[124]]
#> [1] 0.9
#> 
#> [[125]]
#> [1] 0.9
#> 
#> [[126]]
#> [1] 0.2
#> 
#> [[127]]
#> [1] 0.4
#> 
#> [[128]]
#> [1] 0.8
#> 
#> [[129]]
#> [1] 0.1
#> 
#> [[130]]
#> [1] 0.2
#> 
#> [[131]]
#> [1] 0.2
#> 
#> [[132]]
#> [1] 0.7
#> 
#> [[133]]
#> [1] 0.9
#> 
#> [[134]]
#> [1] 0.9
#> 
#> [[135]]
#> [1] 0.6
#> 
#> [[136]]
#> [1] 0.6
#> 
#> [[137]]
#> [1] 0.7
#> 
#> [[138]]
#> [1] 0.2
#> 
#> [[139]]
#> [1] 0
#> 
#> [[140]]
#> [1] 0.9
#> 
#> [[141]]
#> [1] 0.1
#> 
#> [[142]]
#> [1] 1
#> 
#> [[143]]
#> [1] 0.6
#> 
#> [[144]]
#> [1] 0.2
#> 
#> [[145]]
#> [1] 0.9
#> 
#> [[146]]
#> [1] 0
#> 
#> [[147]]
#> [1] 0.6
#> 
#> [[148]]
#> [1] 0.2
#> 
#> [[149]]
#> [1] 0.9
#> 
#> [[150]]
#> [1] 0.8
#> 
#> [[151]]
#> [1] 0.7
#> 
#> [[152]]
#> [1] 0.2
#> 
#> [[153]]
#> [1] 0.7
#> 
#> [[154]]
#> [1] 0.2
#> 
#> [[155]]
#> [1] 0.5
#> 
#> [[156]]
#> [1] 0.3
#> 
#> [[157]]
#> [1] 0.6
#> 
#> [[158]]
#> [1] 0.3
#> 
#> [[159]]
#> [1] 0.5
#> 
#> [[160]]
#> [1] 0.9
#> 
#> [[161]]
#> [1] 1
#> 
#> [[162]]
#> [1] 0.8
#> 
#> [[163]]
#> [1] 0.9
#> 
#> [[164]]
#> [1] 0.6
#> 
#> [[165]]
#> [1] 0.8
#> 
#> [[166]]
#> [1] 1
#> 
#> [[167]]
#> [1] 0.7
#> 
#> [[168]]
#> [1] 0.2
#> 
#> [[169]]
#> [1] 0.3
#> 
#> [[170]]
#> [1] 0.7
#> 
#> [[171]]
#> [1] 0.9
#> 
#> [[172]]
#> [1] 0.2
#> 
#> [[173]]
#> [1] 0
#> 
#> [[174]]
#> [1] 0.9
#> 
#> [[175]]
#> [1] 0.4
#> 
#> [[176]]
#> [1] 0.8
#> 
#> [[177]]
#> [1] 0.7
#> 
#> [[178]]
#> [1] 0.4
#> 
#> [[179]]
#> [1] 0.4
#> 
#> [[180]]
#> [1] 0.7
#> 
#> [[181]]
#> [1] 0.7
#> 
#> [[182]]
#> [1] 0.1
#> 
#> [[183]]
#> [1] 0.9
#> 
#> [[184]]
#> [1] 0.1
#> 
#> [[185]]
#> [1] 0.9
#> 
#> [[186]]
#> [1] 0.1
#> 
#> [[187]]
#> [1] 0.5
#> 
#> [[188]]
#> [1] 0.2
#> 
#> [[189]]
#> [1] 0.7
#> 
#> [[190]]
#> [1] 0.2
#> 
#> [[191]]
#> [1] 1
#> 
#> [[192]]
#> [1] 0.3
#> 
#> [[193]]
#> [1] 0.4
#> 
#> [[194]]
#> [1] 0.9
#> 
#> [[195]]
#> [1] 0.8
#> 
#> [[196]]
#> [1] 0.1
#> 
#> [[197]]
#> [1] 0.9
#> 
#> [[198]]
#> [1] 0.8
#> 
#> [[199]]
#> [1] 0.4
#> 
#> [[200]]
#> [1] 1
#> 
#> [[201]]
#> [1] 0.3
#> 
#> [[202]]
#> [1] 0.9
#> 
#> [[203]]
#> [1] 0.2
#> 
#> [[204]]
#> [1] 0
#> 
#> [[205]]
#> [1] 0.1
#> 
#> [[206]]
#> [1] 0.9
#> 
#> [[207]]
#> [1] 0.2
#> 
#> [[208]]
#> [1] 0.2
#> 
#> [[209]]
#> [1] 0
#> 
#> [[210]]
#> [1] 0.9
#> 
#> [[211]]
#> [1] 0.2
#> 
#> [[212]]
#> [1] 0.8
#> 
#> [[213]]
#> [1] 0.3
#> 
#> [[214]]
#> [1] 0.9
#> 
#> [[215]]
#> [1] 0.3
#> 
#> [[216]]
#> [1] 1
#> 
#> [[217]]
#> [1] 0.6
#> 
#> [[218]]
#> [1] 0.8
#> 
#> [[219]]
#> [1] 0.8
#> 
#> [[220]]
#> [1] 0.8
#> 
#> [[221]]
#> [1] 0.4
#> 
#> [[222]]
#> [1] 0.1
#> 
#> [[223]]
#> [1] 0.1
#> 
#> [[224]]
#> [1] 0.7
#> 
#> [[225]]
#> [1] 0.6
#> 
#> [[226]]
#> [1] 0.7
#> 
#> [[227]]
#> [1] 0.3
#> 
#> [[228]]
#> [1] 1
#> 
#> [[229]]
#> [1] 0.7
#> 
#> [[230]]
#> [1] 0.9
#> 
#> [[231]]
#> [1] 1
#> 
#> [[232]]
#> [1] 0.2
#> 
#> [[233]]
#> [1] 0.7
#> 
#> [[234]]
#> [1] 0.4
#> 
#> [[235]]
#> [1] 0
#> 
#> [[236]]
#> [1] 0.6
#> 
#> [[237]]
#> [1] 0.6
#> 
#> [[238]]
#> [1] 0.4
#> 
#> [[239]]
#> [1] 0.9
#> 
#> [[240]]
#> [1] 0.5
#> 
#> [[241]]
#> [1] 1
#> 
#> [[242]]
#> [1] 0
#> 
#> [[243]]
#> [1] 0.7
#> 
#> [[244]]
#> [1] 0.4
#> 
#> [[245]]
#> [1] 0.9
#> 
#> [[246]]
#> [1] 0.7
#> 
#> [[247]]
#> [1] 0.7
#> 
#> [[248]]
#> [1] 0.8
#> 
#> [[249]]
#> [1] 0.5
#> 
#> [[250]]
#> [1] 0.2
#> 
#> [[251]]
#> [1] 0.5
#> 
#> [[252]]
#> [1] 0.4
#> 
#> [[253]]
#> [1] 0.7
#> 
#> [[254]]
#> [1] 0.6
#> 
#> [[255]]
#> [1] 0.4
#> 
#> [[256]]
#> [1] 1
#> 
#> [[257]]
#> [1] 0.3
#> 
#> [[258]]
#> [1] 0.1
#> 
#> [[259]]
#> [1] 0.6
#> 
#> [[260]]
#> [1] 0.2
#> 
#> [[261]]
#> [1] 0.1
#> 
#> [[262]]
#> [1] 0.9
#> 
#> [[263]]
#> [1] 0.8
#> 
#> [[264]]
#> [1] 0.8
#> 
#> [[265]]
#> [1] 0.5
#> 
#> [[266]]
#> [1] 0.5
#> 
#> [[267]]
#> [1] 0.1
#> 
#> [[268]]
#> [1] 0.4
#> 
#> [[269]]
#> [1] 0.2
#> 
#> [[270]]
#> [1] 0.7
#> 
#> [[271]]
#> [1] 0.7
#> 
#> [[272]]
#> [1] 0.9
#> 
#> [[273]]
#> [1] 0.2
#> 
#> [[274]]
#> [1] 1
#> 
#> [[275]]
#> [1] 0.4
#> 
#> [[276]]
#> [1] 0.7
#> 
#> [[277]]
#> [1] 0.8
#> 
#> [[278]]
#> [1] 0.1
#> 
#> [[279]]
#> [1] 0.5
#> 
#> [[280]]
#> [1] 0.8
#> 
#> [[281]]
#> [1] 0.4
#> 
#> [[282]]
#> [1] 0.6
#> 
#> [[283]]
#> [1] 0.2
#> 
#> [[284]]
#> [1] 0
#> 
#> [[285]]
#> [1] 1
#> 
#> [[286]]
#> [1] 0.2
#> 
#> [[287]]
#> [1] 0.8
#> 
#> [[288]]
#> [1] 0.9
#> 
#> [[289]]
#> [1] 0.8
#> 
#> [[290]]
#> [1] 0.6
#> 
#> [[291]]
#> [1] 0.9
#> 
#> [[292]]
#> [1] 0
#> 
#> [[293]]
#> [1] 0.1
#> 
#> [[294]]
#> [1] 0.1
#> 
#> [[295]]
#> [1] 0.7
#> 
#> [[296]]
#> [1] 0.4
#> 
#> [[297]]
#> [1] 0.1
#> 
#> [[298]]
#> [1] 0.1
#> 
#> [[299]]
#> [1] 0.9
#> 
#> [[300]]
#> [1] 0.2
#> 
#> [[301]]
#> [1] 0.7
#> 
#> [[302]]
#> [1] 0.2
#> 
#> [[303]]
#> [1] 0.5
#> 
#> [[304]]
#> [1] 0.8
#> 
#> [[305]]
#> [1] 0.1
#> 
#> [[306]]
#> [1] 0.3
#> 
#> [[307]]
#> [1] 0.3
#> 
#> [[308]]
#> [1] 0
#> 
#> [[309]]
#> [1] 0.9
#> 
#> [[310]]
#> [1] 0.4
#> 
#> [[311]]
#> [1] 0.8
#> 
#> [[312]]
#> [1] 0.7
#> 
#> [[313]]
#> [1] 0.5
#> 
#> [[314]]
#> [1] 1
#> 
#> [[315]]
#> [1] 0.4
#> 
#> [[316]]
#> [1] 0.2
#> 
#> [[317]]
#> [1] 0.8
#> 
#> [[318]]
#> [1] 0
#> 
#> [[319]]
#> [1] 0.8
#> 
#> [[320]]
#> [1] 0.2
#> 
#> [[321]]
#> [1] 0
#> 
#> [[322]]
#> [1] 0.8
#> 
#> [[323]]
#> [1] 0.8
#> 
#> [[324]]
#> [1] 1
#> 
#> [[325]]
#> [1] 0.4
#> 
#> [[326]]
#> [1] 0.2
#> 
#> [[327]]
#> [1] 0.6
#> 
#> [[328]]
#> [1] 0.8
#> 
#> [[329]]
#> [1] 0
#> 
#> [[330]]
#> [1] 0.7
#> 
#> [[331]]
#> [1] 0.2
#> 
#> [[332]]
#> [1] 0
#> 
#> [[333]]
#> [1] 0.1
#> 
#> [[334]]
#> [1] 0.7
#> 
#> [[335]]
#> [1] 0.6
#> 
#> [[336]]
#> [1] 0.3
#> 
#> [[337]]
#> [1] 0.4
#> 
#> [[338]]
#> [1] 0.7
#> 
#> [[339]]
#> [1] 0.7
#> 
#> [[340]]
#> [1] 0.9
#> 
#> [[341]]
#> [1] 0.5
#> 
#> [[342]]
#> [1] 0.6
#> 
#> [[343]]
#> [1] 0.2
#> 
#> [[344]]
#> [1] 0.6
#> 
#> [[345]]
#> [1] 0.9
#> 
#> [[346]]
#> [1] 0.6
#> 
#> [[347]]
#> [1] 0.8
#> 
#> [[348]]
#> [1] 0.6
#> 
#> [[349]]
#> [1] 0.8
#> 
#> [[350]]
#> [1] 0.4
#> 
#> [[351]]
#> [1] 0.8
#> 
#> [[352]]
#> [1] 0.7
#> 
#> [[353]]
#> [1] 0.3
#> 
#> [[354]]
#> [1] 0.1
#> 
#> [[355]]
#> [1] 0.1
#> 
#> [[356]]
#> [1] 0.8
#> 
#> [[357]]
#> [1] 0.4
#> 
#> [[358]]
#> [1] 0.1
#> 
#> [[359]]
#> [1] 1
#> 
#> [[360]]
#> [1] 0.6
#> 
#> [[361]]
#> [1] 0.1
#> 
#> [[362]]
#> [1] 0.5
#> 
#> [[363]]
#> [1] 0.1
#> 
#> [[364]]
#> [1] 1
#> 
#> [[365]]
#> [1] 0
#> 
#> [[366]]
#> [1] 0.2
#> 
#> [[367]]
#> [1] 0.9
#> 
#> [[368]]
#> [1] 1
#> 
#> [[369]]
#> [1] 0.7
#> 
#> [[370]]
#> [1] 0.9
#> 
#> [[371]]
#> [1] 1
#> 
#> [[372]]
#> [1] 0
#> 
#> [[373]]
#> [1] 0.4
#> 
#> [[374]]
#> [1] 0.6
#> 
#> [[375]]
#> [1] 1
#> 
#> [[376]]
#> [1] 0.3
#> 
#> [[377]]
#> [1] 0.9
#> 
#> [[378]]
#> [1] 0.5
#> 
#> [[379]]
#> [1] 0.9
#> 
#> [[380]]
#> [1] 0.5
#> 
#> [[381]]
#> [1] 0.9
#> 
#> [[382]]
#> [1] 0.9
#> 
#> [[383]]
#> [1] 0.3
#> 
#> [[384]]
#> [1] 0
#>