Get a list of models available from the ollama server

get_list_ollama_models gets a list of models available from the ollama server

Usage

get_list_ollama_models(ollama_connection)

Arguments

ollama_connection

an ollama connection objection object that was created by get_ollama_connection().

Value

A dataframe that contains information about the LLM available from the running ollama server

Details

A simple function to get the models available from the ollama server.

Examples

ollama_conn <- get_ollama_connection()
list_of_models <- get_list_ollama_models(ollama_conn)
#> <error/tibble_error_incompatible_new_cols>
#> Error in `tibble::add_column()`:
#> ! New columns must be compatible with `.data`.
#> ✖ New columns have 2 rows.
#> ℹ `.data` has 6 rows.
#> ---
#> Backtrace:
#>      ▆
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base (local) tryCatchOne(...)
#>   4. │   │ └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base (local) tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─base (local) `<fn>`(...)
#>  12. └─global `<fn>`(...)
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         ├─pkgdown:::unwrap_purrr_error(...)
#>  17.         │ └─base::withCallingHandlers(...)
#>  18.         └─purrr::map(...)
#>  19.           └─purrr:::map_("list", .x, .f, ..., .progress = .progress)
#>  20.             ├─purrr:::with_indexed_errors(...)
#>  21.             │ └─base::withCallingHandlers(...)
#>  22.             ├─purrr:::call_with_cleanup(...)
#>  23.             └─pkgdown (local) .f(.x[[i]], ...)
#>  24.               ├─base::withCallingHandlers(...)
#>  25.               └─pkgdown:::data_reference_topic(...)
#>  26.                 └─pkgdown:::run_examples(...)
#>  27.                   └─pkgdown:::highlight_examples(code, topic, env = env)
#>  28.                     └─downlit::evaluate_and_highlight(...)
#>  29.                       └─evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>  30.                         ├─base::withRestarts(...)
#>  31.                         │ └─base (local) withRestartList(expr, restarts)
#>  32.                         │   ├─base (local) withOneRestart(withRestartList(expr, restarts[-nr]), restarts[[nr]])
#>  33.                         │   │ └─base (local) doWithOneRestart(return(expr), restart)
#>  34.                         │   └─base (local) withRestartList(expr, restarts[-nr])
#>  35.                         │     ├─base (local) withOneRestart(withRestartList(expr, restarts[-nr]), restarts[[nr]])
#>  36.                         │     │ └─base (local) doWithOneRestart(return(expr), restart)
#>  37.                         │     └─base (local) withRestartList(expr, restarts[-nr])
#>  38.                         │       └─base (local) withOneRestart(expr, restarts[[1L]])
#>  39.                         │         └─base (local) doWithOneRestart(return(expr), restart)
#>  40.                         ├─evaluate:::with_handlers(...)
#>  41.                         │ ├─base::eval(call)
#>  42.                         │ │ └─base::eval(call)
#>  43.                         │ └─base::withCallingHandlers(...)
#>  44.                         ├─base::withVisible(eval(expr, envir))
#>  45.                         └─base::eval(expr, envir)
#>  46.                           └─base::eval(expr, envir)
#>  47.                             └─aiworkflow::get_list_ollama_models(ollama_conn)
#>  48.                               ├─base::tryCatch(...)
#>  49.                               │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  50.                               │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  51.                               │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  52.                               └─aiworkflow:::convert_ollama_tags_response_to_tibble(httr2::resp_body_json(result))
#>  53.                                 ├─dplyr::distinct(...)
#>  54.                                 └─tibble::add_column(dplyr::select(temp, -details), details)