Get embeddings for a piece of context through an ollama server instance

get_ollama_embeddings Retrieves the embeddings for a string (can be a series of words, sentence, paragraph, or whole document) through ollama

Usage

get_ollama_embeddings(ollama_connection, model = "bge-large", input)

Arguments

ollama_connection

a connection object that has the information on how to connect to the ollama server

model

The model used to generates the embeddings. Defaults to "all-minilm". You can select any other model available for generating embeddings.

input

A single character vector that contains the text to convert to embeddings

Details

Converts the ollama response for model info to a tibble

Examples

conn <-get_ollama_connection()
embeddings <- get_ollama_embeddings(conn, input="Hi there, how are you doing?")