Get information about one ollama model

get_ollama_model_info Gets information about one model available through ollama.

Usage

get_ollama_model_info(ollama_connection, model)

Arguments

ollama_connection

a connection object that has the information on how to connect to the Oollama server

model

the name of the model that you want to get info from.

Details

Gets information about one model available through ollama.

Examples

#ollama_conn <- get_ollama_connection()
#get_ollama_model_info(ollama_connection= ollama_conn, model="llama3:8b-instruct-q5_0")