Get Qdrant connection

get_qdrant_connection establishes a connection with a qdrant instance. '

Usage

get_qdrant_connection(
  endpoint = "http://localhost",
  port = 6333,
  api_key = NA_character_
)

Arguments

endpoint

the URL pointing to the qdrant instance. Defaults to http://localhost

port

the port to use to connect to the qdrant instance. Defaults to 6333

api_key

optional for security. Defaults to NA. If given, it will use an API key to connect to the Qdrant.

Details

Creates a connection object for qdrant. If you are using the API key to connect to Qdrant, you need to specify it in the api_key parameter.

Examples

conn <- get_qdrant_connection()
#> → Connection to Qdrant confirmed