list the processing skills

list_processing_skills lists the processing skills (i.e. prompts already defined) that you can use out of the box with this package.

Usage

list_processing_skills()

Value

a list of vectors describing the skills available currently in the package.

Details

This gives you a vector of all the skills available to use by default. This can then be used along with the set_processing_skill() function.

Examples

list_processing_skills()
#>  [1] "add_code_comments"              "add_details"                   
#>  [3] "add_emoji"                      "break_down_tasks"              
#>  [5] "categorize_product_complaint"   "categorize_sentiment"          
#>  [7] "chain_of_thought"               "clean_up_audio_transcription"  
#>  [9] "extract_actions"                "extract_names"                 
#> [11] "fix_copy"                       "fix_translation"               
#> [13] "generate_code"                  "generate_creative_introduction"
#> [15] "generate_quiz"                  "identify_product"              
#> [17] "rewrite_active_voice"           "rewrite_blog_copy"             
#> [19] "rewrite_bullet_points"          "rewrite_casual"                
#> [21] "rewrite_emoji"                  "rewrite_in_first_person"       
#> [23] "rewrite_jargon"                 "rewrite_paraphrase"            
#> [25] "rewrite_passive_voice"          "rewrite_positive"              
#> [27] "rewrite_prompt"                 "rewrite_text_as_anonymized"    
#> [29] "tldr"                           "translate"                     
#> [31] "write_abstract"                 "write_article_from_title"      
#> [33] "write_email"                    "write_fiction_chapter"         
#> [35] "write_pitch"                    "write_qa"                      
#> [37] "write_quarto_presentation"      "write_recommendation"          
#> [39] "write_resume_narrative"         "write_short_summary"           
#> [41] "write_text_expansion"           "write_title"                   
#> [43] "write_tweet"